WIP: mm: changes memory map initialization
