+++ arch/x86/mm/Makefile
@@ -1,4 +1,4 @@
-O_OBJS := memory.o memblock.o
+O_OBJS := memory.o memblock.o mmap.o
 O_TARGET := mmar.o
 
 include $(TOPDIR)/Rules.mk
